Fix for Laravel 10
